requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
google-genai==0.2.2
pydub==0.25.1
python-multipart==0.0.20